refactor(frontend): extract ResponseSection and AcceptedSection from PetProfilePage into components/pet-profile; tests passing
